+++ ChangeLog.md
@@ -7,6 +7,7 @@ TODO : pour les options commencer à voir si possible de crééer des tab par el
 
 
 ## Unreleased
+- NEW : Gestion des selects multiple par la fonction stdFormHelper *12/10/2021* - 1.25.0
 - NEW : Gestion des selects par la fonction stdFormHelper *30/09/2021* - 1.24.0
 - NEW : Permissions employé et ajout d'un hook pour la permission d'ajouter un message à un ticket *28/09/2021* - 1.23.0